Display: Base class for widgets with propMacros()
