+++ app/display/model/src/main/java/org/csstudio/display/builder/model/widgets/ArrayWidget.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015-2016 Oak Ridge National Laboratory.
+ * Copyright (c) 2015-2020 Oak Ridge National Laboratory.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,6 @@ package org.csstudio.display.builder.model.widgets;
 
 import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.propBackgroundColor;
 import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.propForegroundColor;
-import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.propMacros;
 import static org.csstudio.display.builder.model.properties.InsetsWidgetProperty.runtimePropInsets;
 
 import java.util.Arrays;
@@ -27,7 +26,6 @@ import org.csstudio.display.builder.model.persist.ModelWriter;
 import org.csstudio.display.builder.model.persist.NamedWidgetColors;
 import org.csstudio.display.builder.model.persist.WidgetColorService;
 import org.csstudio.display.builder.model.properties.WidgetColor;
-import org.phoebus.framework.macros.Macros;
 
 /**
  * An Array Widget contains copies of a child widget. Each copy is assigned the
@@ -36,7 +34,7 @@ import org.phoebus.framework.macros.Macros;
  * @author Amanda Carpenter
  */
 @SuppressWarnings("nls")
-public class ArrayWidget extends PVWidget
+public class ArrayWidget extends MacroWidget
 {
     /** Widget descriptor */
     public static final WidgetDescriptor WIDGET_DESCRIPTOR =
@@ -69,7 +67,6 @@ public class ArrayWidget extends PVWidget
         }
     }
 
-    private volatile WidgetProperty<Macros> macros;
     private volatile ChildrenProperty children;
     private volatile WidgetProperty<WidgetColor> foreground;
     private volatile WidgetProperty<WidgetColor> background;
@@ -84,32 +81,12 @@ public class ArrayWidget extends PVWidget
     protected void defineProperties(final List<WidgetProperty<?>> properties)
     {
         super.defineProperties(properties);
-        properties.add(macros = propMacros.createProperty(this, new Macros()));
         properties.add(children = new ArrayWidgetChildrenProperty(this));
         properties.add(foreground = propForegroundColor.createProperty(this,
                 WidgetColorService.getColor(NamedWidgetColors.TEXT)));
         properties.add(background = propBackgroundColor.createProperty(this,
                 WidgetColorService.getColor(NamedWidgetColors.BACKGROUND)));
         properties.add(insets = runtimePropInsets.createProperty(this, new int[] { 0, 0 }));
-    }
-
-    /**
-     * Array widget extends parent macros
-     *
-     * @return {@link Macros}
-     */
-    @Override
-    public Macros getEffectiveMacros()
-    {
-        final Macros base = super.getEffectiveMacros();
-        final Macros my_macros = propMacros().getValue();
-        return Macros.merge(base, my_macros);
-    }
-
-    /** @return 'macros' property */
-    public WidgetProperty<Macros> propMacros()
-    {
-        return macros;
     }
 
     /** @return Runtime 'children' property*/
