+++ app/display/model/src/main/java/org/csstudio/display/builder/model/widgets/EmbeddedDisplayWidget.java
@@ -9,7 +9,6 @@ package org.csstudio.display.builder.model.widgets;
 
 import static org.csstudio.display.builder.model.ModelPlugin.logger;
 import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.propFile;
-import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.propMacros;
 import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.propTransparent;
 
 import java.util.Arrays;
@@ -34,7 +33,6 @@ import org.csstudio.display.builder.model.persist.XMLTags;
 import org.csstudio.display.builder.model.properties.CommonWidgetProperties;
 import org.csstudio.display.builder.model.properties.EnumWidgetProperty;
 import org.csstudio.display.builder.model.widgets.GroupWidget.Style;
-import org.phoebus.framework.macros.Macros;
 import org.phoebus.framework.persistence.XMLUtil;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
@@ -44,7 +42,7 @@ import org.w3c.dom.Node;
  *  @author Kay Kasemir
  */
 @SuppressWarnings("nls")
-public class EmbeddedDisplayWidget extends VisibleWidget
+public class EmbeddedDisplayWidget extends MacroWidget
 {
     public static final int DEFAULT_WIDTH = 400,
                             DEFAULT_HEIGHT = 300;
@@ -259,7 +257,6 @@ public class EmbeddedDisplayWidget extends VisibleWidget
         }
     }
 
-    private volatile WidgetProperty<Macros> macros;
     private volatile WidgetProperty<String> file;
     private volatile WidgetProperty<Resize> resize;
     private volatile WidgetProperty<String> group_name;
@@ -276,18 +273,11 @@ public class EmbeddedDisplayWidget extends VisibleWidget
     {
         super.defineProperties(properties);
         properties.add(file = propFile.createProperty(this, ""));
-        properties.add(macros = propMacros.createProperty(this, new Macros()));
         properties.add(resize = propResize.createProperty(this, Resize.None));
         properties.add(group_name = propGroupName.createProperty(this, ""));
         properties.add(embedded_model = runtimeModel.createProperty(this, null));
         properties.add(transparent = propTransparent.createProperty(this, false));
         BorderSupport.addBorderProperties(this, properties);
-    }
-
-    /** @return 'macros' property */
-    public WidgetProperty<Macros> propMacros()
-    {
-        return macros;
     }
 
     /** @return 'file' property */
@@ -325,16 +315,5 @@ public class EmbeddedDisplayWidget extends VisibleWidget
             throws Exception
     {
         return new EmbeddedDisplayWidgetConfigurator(persisted_version);
-    }
-
-    /** Embedded widget adds/replaces parent macros
-     *  @return {@link Macros}
-     */
-    @Override
-    public Macros getEffectiveMacros()
-    {
-        final Macros base = super.getEffectiveMacros();
-        final Macros my_macros = propMacros().getValue();
-        return base == null ? my_macros : Macros.merge(base, my_macros);
     }
 }
